chore: break up long tests file
- due to pylint `C0302(too-many-lines)`
