+++ tests/unit/gaze/gaze_test.py
@@ -20,7 +20,6 @@
 """Test all Gaze functionality."""
 from __future__ import annotations
 
-import os
 from copy import deepcopy
 
 import numpy as np
@@ -1801,368 +1800,3 @@ def _create_gaze():
             ),
         ),
     )
-
-
-def test_gaze_save_csv(tmp_path):
-
-    gaze = _create_gaze()
-    # Saving Gaze to tmp_path
-    gaze.save(
-        dirpath=tmp_path,
-        verbose=2,
-        extension='csv',
-    )
-    assert os.path.exists(tmp_path / 'samples.csv')
-    assert os.path.exists(tmp_path / 'events.csv')
-    assert os.path.exists(tmp_path / 'experiment.yaml')
-
-
-def test_gaze_save_feather(tmp_path):
-    gaze = _create_gaze()
-    # Saving Gaze to tmp_path
-    gaze.save(
-        dirpath=tmp_path,
-        verbose=2,
-        extension='feather',
-    )
-    assert os.path.exists(tmp_path / 'samples.feather')
-    assert os.path.exists(tmp_path / 'events.feather')
-    assert os.path.exists(tmp_path / 'experiment.yaml')
-
-
-def test_gaze_save_without_events(tmp_path):
-
-    gaze = _create_gaze()
-
-    # Saving Gaze to tmp_path
-    gaze.save(
-        dirpath=tmp_path,
-        save_events=False,
-        verbose=2,
-        extension='csv',
-    )
-    assert not os.path.exists(tmp_path / 'events.csv')
-    assert os.path.exists(tmp_path / 'samples.csv')
-    assert os.path.exists(tmp_path / 'experiment.yaml')
-
-
-def test_gaze_save_without_samples(tmp_path):
-
-    gaze = _create_gaze()
-
-    # Saving Gaze to tmp_path
-    gaze.save(
-        dirpath=tmp_path,
-        save_samples=False,
-        verbose=2,
-        extension='csv',
-    )
-    assert os.path.exists(tmp_path / 'events.csv')
-    assert not os.path.exists(tmp_path / 'samples.csv')
-    assert os.path.exists(tmp_path / 'experiment.yaml')
-
-
-def test_gaze_save_without_experiment(tmp_path):
-
-    gaze = _create_gaze()
-
-    # Saving Gaze to tmp_path
-    gaze.save(
-        dirpath=tmp_path,
-        save_experiment=False,
-        verbose=1,
-        extension='csv',
-    )
-    assert os.path.exists(tmp_path / 'events.csv')
-    assert os.path.exists(tmp_path / 'samples.csv')
-    assert not os.path.exists(tmp_path / 'experiment.yaml')
-
-
-def test_gaze_save_with_empty_events(tmp_path):
-
-    gaze = _create_gaze()
-    gaze.events = None
-
-    with pytest.raises(ValueError):
-        gaze.save(
-            dirpath=tmp_path,
-            save_events=True,
-            verbose=2,
-            extension='csv',
-        )
-
-
-def test_gaze_save_wrong_extension_events(tmp_path):
-    gaze = _create_gaze()
-
-    with pytest.raises(ValueError):
-        gaze.save(
-            dirpath=tmp_path,
-            verbose=0,
-            extension='blabla',
-        )
-
-
-def test_gaze_save_wrong_extension_samples(tmp_path):
-    gaze = _create_gaze()
-
-    with pytest.raises(ValueError):
-        gaze.save(
-            dirpath=tmp_path,
-            save_events=False,
-            verbose=1,
-            extension='blabla',
-        )
-
-
-def test_gaze_save_empty_experiment(tmp_path):
-    gaze = _create_gaze()
-    gaze.experiment = None
-
-    gaze.save(
-        dirpath=tmp_path,
-        verbose=1,
-        extension='csv',
-    )
-    assert os.path.exists(tmp_path / 'events.csv')
-    assert os.path.exists(tmp_path / 'samples.csv')
-    assert not os.path.exists(tmp_path / 'experiment.yaml')
-
-
-def test_gaze_save_empty_experiment_true_save(tmp_path):
-    gaze = _create_gaze()
-    gaze.experiment = None
-
-    with pytest.raises(ValueError):
-        gaze.save(
-            dirpath=tmp_path,
-            save_experiment=True,
-            verbose=1,
-            extension='csv',
-        )
-
-
-def test_gaze_save_samples_csv_no_warning_without_nested_columns(tmp_path):
-    samples = pl.DataFrame(
-        {'x': [0, 1], 'y': [1, 0], 'trial': [1, 2]},
-        schema={'x': pl.Float64, 'y': pl.Float64, 'trial': pl.Int64},
-    )
-    gaze = Gaze(samples=samples, pixel_columns=['x', 'y'], trial_columns='trial')
-
-    gaze.save(
-        dirpath=tmp_path,
-        save_samples=True,
-        save_events=False,
-        save_experiment=False,
-        verbose=1,
-        extension='csv',
-    )
-
-
-@pytest.mark.parametrize(
-    'gaze,expected_file',
-    [
-        pytest.param(
-            Gaze(
-                pl.DataFrame({'x': [1, 2], 'y': [3, 4]}),
-                pixel_columns=['x', 'y'],
-                metadata={'key': 'value', 'nested': {'inner': 42}},
-            ),
-            'metadata.yaml',
-            id='with_metadata',
-        ),
-    ],
-)
-def test_gaze_save_metadata(tmp_path, gaze, expected_file):
-    gaze.save(tmp_path, save_metadata=True, verbose=2)
-    assert os.path.exists(tmp_path / expected_file)
-
-
-@pytest.mark.parametrize(
-    'gaze,save_func,expected_file,raises_error,error_match',
-    [
-        pytest.param(
-            Gaze(
-                pl.DataFrame({'x': [1, 2], 'y': [3, 4]}),
-                pixel_columns=['x', 'y'],
-                messages=pl.DataFrame({'time': [0], 'content': ['msg']}),
-            ),
-            lambda g, p: g.save_messages(p / 'messages.feather', verbose=2),
-            'messages.feather',
-            False,
-            None,
-            id='messages_with_data',
-        ),
-        pytest.param(
-            Gaze(pl.DataFrame({'x': [1, 2], 'y': [3, 4]}), pixel_columns=['x', 'y']),
-            lambda g, p: g.save_messages(p / 'messages.feather', verbose=2),
-            None,
-            True,
-            'No messages',
-            id='messages_none',
-        ),
-        pytest.param(
-            Gaze(
-                pl.DataFrame({'x': [1, 2], 'y': [3, 4]}),
-                pixel_columns=['x', 'y'],
-                calibrations=pl.DataFrame({'timestamp': [0], 'num_points': [9]}),
-            ),
-            lambda g, p: g.save_calibrations(p / 'calibrations.feather', verbose=2),
-            'calibrations.feather',
-            False,
-            None,
-            id='calibrations_with_data',
-        ),
-        pytest.param(
-            Gaze(pl.DataFrame({'x': [1, 2], 'y': [3, 4]}), pixel_columns=['x', 'y']),
-            lambda g, p: g.save_calibrations(p / 'calibrations.feather', verbose=2),
-            None,
-            True,
-            'No calibrations',
-            id='calibrations_none',
-        ),
-        pytest.param(
-            Gaze(
-                pl.DataFrame({'x': [1, 2], 'y': [3, 4]}),
-                pixel_columns=['x', 'y'],
-                validations=pl.DataFrame({'timestamp': [0], 'accuracy_avg': [0.5]}),
-            ),
-            lambda g, p: g.save_validations(p / 'validations.feather', verbose=2),
-            'validations.feather',
-            False,
-            None,
-            id='validations_with_data',
-        ),
-        pytest.param(
-            Gaze(pl.DataFrame({'x': [1, 2], 'y': [3, 4]}), pixel_columns=['x', 'y']),
-            lambda g, p: g.save_validations(p / 'validations.feather', verbose=2),
-            None,
-            True,
-            'No validations',
-            id='validations_none',
-        ),
-    ],
-)
-def test_gaze_save_dataframes(
-    tmp_path,
-    gaze,
-    save_func,
-    expected_file,
-    raises_error,
-    error_match,
-):
-    if raises_error:
-        with pytest.raises(ValueError, match=error_match):
-            save_func(gaze, tmp_path)
-    else:
-        save_func(gaze, tmp_path)
-        assert os.path.exists(tmp_path / expected_file)
-
-
-@pytest.mark.parametrize(
-    'save_method',
-    [
-        pytest.param('save_messages', id='messages'),
-        pytest.param('save_calibrations', id='calibrations'),
-        pytest.param('save_validations', id='validations'),
-    ],
-)
-def test_gaze_save_wrong_extension(tmp_path, save_method):
-    gaze = Gaze(pl.DataFrame({'x': [1, 2], 'y': [3, 4]}), pixel_columns=['x', 'y'])
-    if save_method == 'save_messages':
-        gaze.messages = pl.DataFrame({'time': [0], 'content': ['msg']})
-    elif save_method == 'save_calibrations':
-        gaze.calibrations = pl.DataFrame({'timestamp': [0], 'num_points': [9]})
-    else:
-        gaze.validations = pl.DataFrame({'timestamp': [0], 'accuracy_avg': [0.5]})
-
-    with pytest.raises(ValueError, match='unsupported file format'):
-        getattr(gaze, save_method)(tmp_path / 'test.txt')
-
-
-def test_transform_early_return_on_empty_grouped_frames():
-    # Create an empty samples frame with only the trial column so grouping yields no groups
-    samples = pl.DataFrame(schema={'trial': pl.Int64})
-    gaze = Gaze(samples=samples, trial_columns='trial')
-
-    # Calling a transform that would normally require an input column should do nothing
-    # because grouped_frames will be empty and the method returns early.
-    before = gaze.samples.clone()
-    gaze.clip(lower_bound=None, upper_bound=None, input_column='position', output_column='clipped')
-    after = gaze.samples
-
-    # Ensure samples are unchanged (no new columns, still empty)
-    assert before.schema == after.schema
-    assert before.shape == after.shape
-
-
-def test_transform_returns_early_when_groupby_yields_no_groups(monkeypatch):
-    # Create a non-empty samples DataFrame with a trial column so is_empty() is False
-    samples = pl.DataFrame({'trial': [1]})
-    # Creating a Gaze without identifiable components emits a UserWarning
-    with pytest.warns(UserWarning, match='no components could be inferred'):
-        gaze = Gaze(samples=samples, trial_columns='trial')
-
-    # Define a dummy transform callable that does not require n_components or specific columns
-    def dummy_transform(**_kwargs):  # pragma: no cover - exercised via transform
-        return pl.lit(1).alias('dummy')
-
-    # Monkeypatch polars.DataFrame.group_by to return an empty iterator, simulating no groups
-    def fake_group_by(self, keys, maintain_order=True):  # pylint: disable=unused-argument
-        return iter([])
-
-    monkeypatch.setattr(pl.DataFrame, 'group_by', fake_group_by, raising=True)
-
-    before = gaze.samples.clone()
-    # Invoke transform - due to patched group_by producing no groups, it should early-return
-    gaze.transform(dummy_transform)
-    after = gaze.samples
-
-    # Ensure samples are unchanged (no new columns added)
-    assert before.schema == after.schema
-    assert before.shape == after.shape
-
-
-@pytest.mark.parametrize(
-    'trials',
-    [
-        pytest.param([1], id='single_row_single_group'),
-        pytest.param([1, 1, 2], id='multiple_rows_multiple_groups'),
-    ],
-)
-def test_transform_grouped_path_non_empty_samples(trials):
-    # Non-empty samples so the is_empty() guard is False - ensure grouping yields groups
-    samples = pl.DataFrame({'trial': trials})
-
-    # Creating a Gaze without identifiable components emits a UserWarning
-    with pytest.warns(UserWarning, match='no components could be inferred'):
-        gaze = Gaze(samples=samples, trial_columns='trial')
-
-    # Define a simple transform that doesn't require n_components/columns
-    def dummy_transform(**_kwargs):  # pragma: no cover - exercised via transform
-        return pl.lit(7).alias('dummy')
-
-    gaze.transform(dummy_transform)
-
-    # Verify that the transform was applied through the grouped path
-    assert 'dummy' in gaze.samples.columns
-    assert gaze.samples['dummy'].to_list() == [7] * len(trials)
-
-
-def test_transform_grouped_path_empty_samples_early_return():
-    # Empty samples with a trial column: should reach the grouped-path empty check
-    # and return early there (not the earlier n_components guard), so we use a dummy
-    # transform that does not require n_components.
-    samples = pl.DataFrame(schema={'trial': pl.Int64})
-    gaze = Gaze(samples=samples, trial_columns='trial')
-
-    def dummy_transform(**_kwargs):  # pragma: no cover - exercised via transform
-        return pl.lit(1).alias('dummy')
-
-    before = gaze.samples.clone()
-    gaze.transform(dummy_transform)
-    after = gaze.samples
-
-    # Ensure samples are unchanged (no new columns, still empty)
-    assert before.schema == after.schema
-    assert before.shape == after.shape
